+++ RELEASE-NOTES.md
@@ -2,6 +2,7 @@
 
 * #14 Allow multiple routing keys for a combination of exhcnage and queue
 * #15 Make RmqConnectionManager injectable
+* #16 Injectable RmqTopologyManager
 
 ## 2.0.RC1